chore: mobile topbar component
